requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.18
django-elasticsearch-dsl==7.3
django-extensions==3.2.1
djangorestframework==3.12.4
docopt==0.6.2
elastic-transport==8.4.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
favicon==0.7.0
h11==0.14.0
idna==3.4
MarkupPy==1.14
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.1
outcome==1.2.0
pandas==1.5.3
pipreqs==0.4.11
plotly==5.13.0
PySocks==1.7.1
pysolr==3.9.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
selenium==4.8.3
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.3.0
tenacity==8.2.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2022.7
urllib3==1.26.14
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9